JENKINS-34135 Make JiraTriggerExecutor to work without job read permission.
The implementation of JiraTriggerExecutor was previously querying Jenkins over and over again. Instead of doing the search, we are now keeping track of the created trigger object every time they are created in Descriptor, hence JiraTriggerExecutor is now just grabbing the list of triggers and execute the run method. This works similarly in org.jenkinsci.plugins.ghprb.GhprbTrigger.